refactor: split ShareableStatsVideo into focused pieces
Carve out:
- stats-video-helpers.ts — pure helpers (status text, mime picker,
  filename builder, days-without-wedgie)
- stats-video-canvas.ts — all canvas drawing (frame layouts, waves,
  confetti, logo badge, image loader)
- useStatsVideo.ts — recorder hook (state + generateVideo + download
  + reset), with constants for dimensions, FPS, bitrate, duration
- StatsVideoFormatToggle, StatsVideoResult — UI sections

The component itself drops from ~790 to ~120 lines. Helpers get
unit tests; canvas drawing remains hard to assert meaningfully so
isn't covered.
