+++ src/components/home/ShareableStatsVideo.tsx
@@ -1,4 +1,5 @@
-import { useState, useRef } from "react";
+"use client";
+
 import { Loader } from "../shared/Loader";
 import {
   Dialog,
@@ -6,650 +7,34 @@ import {
   DialogTrigger,
   DialogTitle,
 } from "~/components/ui/dialog";
-import {
-  MonitorDown,
-  RectangleVertical,
-  RectangleHorizontal,
-} from "lucide-react";
+import { MonitorDown } from "lucide-react";
 import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
 } from "~/components/ui/tooltip";
+import type { ShareableStats } from "./stats-video-helpers";
+import { useStatsVideo } from "./useStatsVideo";
+import { StatsVideoFormatToggle } from "./StatsVideoFormatToggle";
+import { StatsVideoResult } from "./StatsVideoResult";
 
 interface ShareableStatsVideoProps {
-  stats: {
-    totalWedgies: number;
-    currentPace: number;
-    gamesPlayed: number;
-    previousRecord: number;
-    lastWedgie: Date | string | null;
-    liveGames: boolean;
-  };
-}
-
-interface WaveLayer {
-  speedMultiplier: number;
-  yOffset: number;
-  opacity: number;
-  phase: number;
-}
-
-interface ConfettiParticle {
-  x: number;
-  y: number;
-  vx: number;
-  vy: number;
-  radius: number;
-  color: string;
-  opacity: number;
-  gravity: number;
-  drag: number;
-}
-
-function getWedgieDays(lastWedgie: Date | string | null) {
-  if (!lastWedgie) return { daysWithoutWedgie: 0, hasNewWedgie: false };
-  const elapsedMs = new Date().getTime() - new Date(lastWedgie).getTime();
-  const elapsedDays = elapsedMs / (1000 * 60 * 60 * 24);
-  return {
-    daysWithoutWedgie: Math.ceil(elapsedDays) - 1,
-    hasNewWedgie: elapsedDays < 1,
-  };
+  stats: ShareableStats;
 }
 
 export function ShareableStatsVideo({ stats }: ShareableStatsVideoProps) {
-  const [isGenerating, setIsGenerating] = useState(false);
-  const [videoBlob, setVideoBlob] = useState<Blob | null>(null);
-  const [progress, setProgress] = useState(0);
-  const [videoType, setVideoType] = useState<"desktop" | "mobile">("desktop");
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-  const imageRef = useRef<HTMLImageElement | null>(null); // Ref to store the logo image
-
-  // Initialize wave layers with increased yOffset for better visibility
-  const waveLayers: WaveLayer[] = [
-    { speedMultiplier: 0.25, yOffset: -15, opacity: 0.3, phase: 0 },
-    { speedMultiplier: 0.3, yOffset: -10, opacity: 0.5, phase: Math.PI / 2 },
-    { speedMultiplier: 0.35, yOffset: 10, opacity: 0.7, phase: Math.PI },
-    {
-      speedMultiplier: 0.4,
-      yOffset: 0,
-      opacity: 1.0,
-      phase: (3 * Math.PI) / 2,
-    },
-  ];
-
-  const confettiColors = [
-    "#eaff00",
-    "#ff03ff",
-    "#180138",
-    "#542299",
-    "#efff40",
-  ];
-  const confettiParticlesRef = useRef<ConfettiParticle[]>([]);
-
-  const createConfettiParticles = (
-    width: number,
-    height: number,
-  ): ConfettiParticle[] => {
-    const particles: ConfettiParticle[] = [];
-    for (let i = 0; i < 200; i++) {
-      particles.push({
-        x: Math.random() * width,
-        y: Math.random() * height,
-        vx: (Math.random() - 0.5) * 4,
-        vy: Math.random() * -3 - 1,
-        radius: Math.random() * 5 + 3,
-        color:
-          confettiColors[Math.floor(Math.random() * confettiColors.length)]!,
-        opacity: Math.random() * 0.6 + 0.4,
-        gravity: 0.08,
-        drag: 0.98,
-      });
-    }
-    return particles;
-  };
-
-  const drawConfetti = (
-    ctx: CanvasRenderingContext2D,
-    width: number,
-    height: number,
-  ) => {
-    const particles = confettiParticlesRef.current;
-    for (const p of particles) {
-      p.vy += p.gravity;
-      p.vx *= p.drag;
-      p.vy *= p.drag;
-      p.x += p.vx;
-      p.y += p.vy;
-
-      // Recycle particles that fall off-screen
-      if (p.y > height + p.radius) {
-        p.y = -p.radius;
-        p.x = Math.random() * width;
-        p.vy = Math.random() * -1;
-        p.vx = (Math.random() - 0.5) * 4;
-      }
-
-      ctx.beginPath();
-      ctx.arc(p.x, p.y, p.radius, 0, Math.PI * 2);
-      ctx.fillStyle = p.color;
-      ctx.globalAlpha = p.opacity;
-      ctx.fill();
-    }
-    ctx.globalAlpha = 1;
-  };
-
-  /**
-   * Helper function to load an image.
-   * @param url - The URL of the image to load.
-   * @returns A promise that resolves to the loaded HTMLImageElement.
-   */
-  const loadImage = (url: string): Promise<HTMLImageElement> => {
-    return new Promise((resolve, reject) => {
-      const img = new Image();
-      img.crossOrigin = "Anonymous";
-      img.src = url;
-      img.onload = () => resolve(img);
-      img.onerror = () => reject(new Error(`Failed to load image: ${url}`));
-    });
-  };
-
-  /**
-   * Draws a single wave layer on the canvas.
-   * @param ctx - The canvas rendering context.
-   * @param layer - The wave layer configuration.
-   * @param waveLength - The wavelength of the wave.
-   * @param baseY - The base Y position for the wave.
-   */
-  const drawWaveLayer = (
-    ctx: CanvasRenderingContext2D,
-    layer: WaveLayer,
-    waveLength: number,
-    baseY: number,
-  ) => {
-    // Update the phase for animation
-    layer.phase += layer.speedMultiplier * 0.02; // Adjust frameSpeed here if needed
-    if (layer.phase > 2 * Math.PI) {
-      layer.phase -= 2 * Math.PI;
-    }
-
-    // Draw two waves per layer for seamless looping
-    for (let i = 0; i < 2; i++) {
-      ctx.beginPath();
-      ctx.moveTo(0, canvasRef.current!.height); // Start from bottom-left
-
-      for (let x = 0; x <= canvasRef.current!.width + 10; x += 10) {
-        const dx = (x / waveLength) * 2 * Math.PI + layer.phase + i * Math.PI;
-        const yOffsetCalculated = Math.sin(dx) * 40; // Wave amplitude
-        ctx.lineTo(x, baseY + layer.yOffset + yOffsetCalculated);
-      }
-
-      ctx.lineTo(canvasRef.current!.width, canvasRef.current!.height); // Bottom-right
-      ctx.lineTo(0, canvasRef.current!.height); // Back to bottom-left
-      ctx.closePath();
-
-      ctx.fillStyle = `rgba(255, 0, 255, ${layer.opacity})`; // Pink with variable opacity
-      ctx.fill();
-    }
-  };
-
-  const drawLogoAndUrlBadge = (
-    ctx: CanvasRenderingContext2D,
-    width: number,
-    height: number,
-    logoY: number,
-    urlBottomOffset: number,
-  ) => {
-    if (imageRef.current) {
-      const logoWidth = 300;
-      const logoHeight = 100;
-      const logoX = width / 2 - logoWidth / 2;
-
-      ctx.fillStyle = "rgba(31, 0, 77, 0.75)";
-      ctx.beginPath();
-      ctx.roundRect(
-        logoX - 60,
-        logoY - 60,
-        logoWidth + 120,
-        logoHeight + 120,
-        20,
-      );
-      ctx.fill();
-      ctx.drawImage(imageRef.current, logoX, logoY, logoWidth, logoHeight);
-    }
-
-    const urlText = "WEDGIETRACKER.COM";
-    ctx.font = "700 24px Inter";
-    const textMetrics = ctx.measureText(urlText);
-    const padding = 40;
-    const buttonWidth = textMetrics.width + padding * 2;
-    const buttonHeight = 50;
-    const buttonX = width / 2 - buttonWidth / 2;
-    const buttonY = height - urlBottomOffset;
-
-    ctx.strokeStyle = "rgb(23,0,43)";
-    ctx.lineWidth = 2;
-    ctx.beginPath();
-    ctx.roundRect(
-      buttonX,
-      buttonY,
-      buttonWidth,
-      buttonHeight,
-      buttonHeight / 2,
-    );
-    ctx.stroke();
-
-    ctx.fillStyle = "rgb(23,0,43)";
-    ctx.textAlign = "center";
-    ctx.fillText(urlText, width / 2, buttonY + 34);
-  };
-
-  /**
-   * Common preamble: clears canvas, draws waves and confetti, returns
-   * shared animation values used by every frame layout.
-   */
-  const setupFrame = (
-    ctx: CanvasRenderingContext2D,
-    progress: number,
-    stats: ShareableStatsVideoProps["stats"],
-  ) => {
-    const canvas = ctx.canvas;
-    const width = canvas.width;
-    const height = canvas.height;
-
-    const numberAnimationProgress = Math.min(progress * 20, 1);
-    const currentTotalWedgies = Math.round(
-      stats.totalWedgies * numberAnimationProgress,
-    );
-    const currentPaceValue = Math.round(
-      stats.currentPace * numberAnimationProgress,
-    );
-    const currentGamesValue = Math.round(
-      stats.gamesPlayed * numberAnimationProgress,
-    );
-
-    const fillPercentage = Math.min((stats.totalWedgies / 50) * 100, 100);
-    const currentFillPercentage = Math.min(
-      fillPercentage,
-      ((progress * 20) / 2) * 100,
-    );
-
-    ctx.fillStyle = "rgb(23,0,43)";
-    ctx.fillRect(0, 0, width, height);
-
-    const baseY = height - (currentFillPercentage / 100) * height;
-
-    waveLayers.forEach((layer) => {
-      drawWaveLayer(ctx, layer, width * 0.4, baseY);
-    });
-
-    if (fillPercentage >= 100 && currentFillPercentage >= fillPercentage) {
-      drawConfetti(ctx, width, height);
-    }
-
-    return {
-      width,
-      height,
-      currentTotalWedgies,
-      currentPaceValue,
-      currentGamesValue,
-    };
-  };
-
-  /**
-   * Draws the current frame on the canvas.
-   * @param ctx - The canvas rendering context.
-   * @param progress - The progress of the video generation (0 to 1).
-   * @param stats - The stats data.
-   */
-  const drawFrame = (
-    ctx: CanvasRenderingContext2D,
-    progress: number,
-    stats: ShareableStatsVideoProps["stats"],
-  ) => {
-    const {
-      width,
-      height,
-      currentTotalWedgies,
-      currentPaceValue,
-      currentGamesValue,
-    } = setupFrame(ctx, progress, stats);
-
-    // Stats container dimensions
-    const statsWidth = width * 0.3;
-    const statsHeight = height * 0.4;
-    const statsX = (width - statsWidth) / 2;
-    const statsY = (height - statsHeight) / 2 + 60;
-
-    // Add semi-transparent background for stats with rounded corners
-    ctx.fillStyle = "rgba(31, 0, 77, 0.75)";
-    ctx.beginPath();
-    ctx.roundRect(statsX, statsY, statsWidth, statsHeight, 20);
-    ctx.fill();
-
-    // Draw status text (WE'RE AT / NEW ALL-TIME RECORD / ALL-TIME RECORD TIED)
-    const statusText =
-      stats.totalWedgies > stats.previousRecord
-        ? "NEW ALL-TIME RECORD"
-        : stats.totalWedgies === stats.previousRecord
-          ? "ALL-TIME RECORD TIED"
-          : "WE'RE AT";
-    ctx.font = "700 32px Inter";
-    ctx.fillStyle = "#EAFF00"; // Yellow
-    ctx.textAlign = "center";
-    ctx.fillText(statusText, width / 2, statsY + 60);
-
-    // Draw total wedgies number (animated) with glow
-    ctx.font = "900 280px Inter";
-    ctx.textAlign = "center";
-    ctx.shadowColor = "rgba(23, 0, 43, 0.8)";
-    ctx.shadowBlur = 30;
-    ctx.shadowOffsetX = 0;
-    ctx.shadowOffsetY = 0;
-    ctx.fillStyle = "#EAFF00"; // Yellow
-    ctx.fillText(
-      currentTotalWedgies.toString(),
-      width / 2,
-      statsY + statsHeight / 2 + 85,
-    );
-    ctx.shadowBlur = 0;
-
-    // Draw "WEDGIES" text
-    ctx.font = "900 75px Inter";
-    ctx.fillText("WEDGIES", width / 2, statsY + statsHeight - 35);
-
-    // Add rounded rectangles for bottom stats
-    const statBoxWidth = width * 0.15;
-    const statBoxHeight = 200;
-    const leftStatX = width * 0.25 - statBoxWidth / 2;
-    const rightStatX = width * 0.75 - statBoxWidth / 2;
-
-    // Bottom stats section
-    const bottomY = (height - statsHeight) / 2 + 180;
-
-    const showPace = stats.currentPace !== stats.totalWedgies;
-
-    // Background boxes for stats
-    ctx.fillStyle = "rgba(31, 0, 77, 0.75)";
-    ctx.beginPath();
-    ctx.roundRect(leftStatX, bottomY, statBoxWidth, statBoxHeight, 20);
-    ctx.roundRect(rightStatX, bottomY, statBoxWidth, statBoxHeight, 20);
-    ctx.fill();
-
-    // Draw Pace or Games Played (left side) - animated number
-    if (showPace) {
-      ctx.font = "900 60px Inter";
-      ctx.fillStyle = "#FF00FF"; // Pink
-      ctx.textAlign = "center";
-      ctx.fillText("PACE", width * 0.25, bottomY + 180);
-      ctx.font = "900 120px Inter";
-      ctx.fillStyle = "#EAFF00"; // Yellow
-      ctx.fillText(currentPaceValue.toString(), width * 0.25, bottomY + 120);
-    } else {
-      ctx.font = "900 120px Inter";
-      ctx.fillStyle = "#EAFF00"; // Yellow
-      ctx.textAlign = "center";
-      ctx.fillText(currentGamesValue.toString(), width * 0.25, bottomY + 116);
-      ctx.font = "700 28px Inter";
-      ctx.fillStyle = "#FF00FF"; // Pink
-      ctx.fillText("GAMES", width * 0.25, bottomY + 152);
-      ctx.fillText("PLAYED", width * 0.25, bottomY + 182);
-    }
-
-    // Draw days without wedgies or new wedgie (right side)
-    const { daysWithoutWedgie, hasNewWedgie } = getWedgieDays(stats.lastWedgie);
-
-    if (hasNewWedgie) {
-      ctx.font = "900 64px Inter";
-      ctx.fillStyle = "#FF00FF"; // Pink
-      ctx.fillText("NEW", width * 0.75, bottomY + 110);
-      ctx.font = "900 50px Inter";
-      ctx.fillStyle = "#EAFF00"; // Yellow
-      ctx.fillText("WEDGIE", width * 0.75, bottomY + 150);
-    } else {
-      ctx.font = "900 120px Inter";
-      ctx.fillStyle = "#EAFF00"; // Yellow
-      ctx.fillText(daysWithoutWedgie.toString(), width * 0.75, bottomY + 116);
-      ctx.font = "700 28px Inter";
-      ctx.fillStyle = "#FF00FF"; // Pink
-      ctx.fillText("DAYS WITHOUT", width * 0.75, bottomY + 152);
-      ctx.fillText("WEDGIES", width * 0.75, bottomY + 182);
-    }
-
-    drawLogoAndUrlBadge(ctx, width, height, 150, 159);
-  };
-
-  const drawVerticalFrame = (
-    ctx: CanvasRenderingContext2D,
-    progress: number,
-    stats: ShareableStatsVideoProps["stats"],
-  ) => {
-    const {
-      width,
-      height,
-      currentTotalWedgies,
-      currentPaceValue,
-      currentGamesValue,
-    } = setupFrame(ctx, progress, stats);
-
-    // Stats container dimensions
-    const statsWidth = width * 0.65;
-    const statsHeight = height * 0.3;
-    const statsX = (width - statsWidth) / 2;
-    const statsY = (height - statsHeight) / 2 - 150;
-
-    // Add semi-transparent background for stats with rounded corners
-    ctx.fillStyle = "rgba(31, 0, 77, 0.75)";
-    ctx.beginPath();
-    ctx.roundRect(statsX, statsY, statsWidth, statsHeight, 20);
-    ctx.fill();
-
-    // Draw status text (WE'RE AT / NEW ALL-TIME RECORD / ALL-TIME RECORD TIED)
-    const statusText =
-      stats.totalWedgies > stats.previousRecord
-        ? "NEW ALL-TIME RECORD"
-        : stats.totalWedgies === stats.previousRecord
-          ? "ALL-TIME RECORD TIED"
-          : "WE'RE AT";
-    ctx.font = "700 32px Inter";
-    ctx.fillStyle = "#EAFF00"; // Yellow
-    ctx.textAlign = "center";
-    ctx.fillText(statusText, width / 2, statsY + 85);
-
-    // Draw total wedgies number (animated) with glow
-    ctx.font = "900 380px Inter";
-    ctx.textAlign = "center";
-    ctx.shadowColor = "rgba(23, 0, 43, 0.8)";
-    ctx.shadowBlur = 30;
-    ctx.shadowOffsetX = 0;
-    ctx.shadowOffsetY = 0;
-    ctx.fillStyle = "#EAFF00"; // Yellow
-    ctx.fillText(
-      currentTotalWedgies.toString(),
-      width / 2,
-      statsY + statsHeight / 2 + 125,
-    );
-    ctx.shadowBlur = 0;
-
-    // Draw "WEDGIES" text
-    ctx.font = "900 75px Inter";
-    ctx.fillText("WEDGIES", width / 2, statsY + statsHeight - 55);
-
-    // Add rounded rectangles for bottom stats
-    const statBoxWidth = width * 0.3;
-    const statBoxHeight = 250;
-    const leftStatX = width * 0.68 - statBoxWidth / 2;
-    const rightStatX = width * 0.32 - statBoxWidth / 2;
-
-    // Bottom stats section
-    const bottomY = (height - statsHeight) / 2 + 550;
-
-    const showPace = stats.currentPace !== stats.totalWedgies;
-
-    // Background boxes for stats
-    ctx.fillStyle = "rgba(31, 0, 77, 0.75)";
-    ctx.beginPath();
-    ctx.roundRect(rightStatX, bottomY, statBoxWidth, statBoxHeight, 20);
-    ctx.roundRect(leftStatX, bottomY, statBoxWidth, statBoxHeight, 20);
-    ctx.fill();
-
-    // Draw Pace or Games Played (left side) - animated number
-    if (showPace) {
-      ctx.font = "900 90px Inter";
-      ctx.fillStyle = "#FF00FF"; // Pink
-      ctx.textAlign = "center";
-      ctx.fillText("PACE", width * 0.32, bottomY + 215);
-      ctx.font = "900 165px Inter";
-      ctx.fillStyle = "#EAFF00"; // Yellow
-      ctx.fillText(currentPaceValue.toString(), width * 0.32, bottomY + 155);
-    } else {
-      ctx.font = "900 140px Inter";
-      ctx.fillStyle = "#EAFF00"; // Yellow
-      ctx.textAlign = "center";
-      ctx.fillText(currentGamesValue.toString(), width * 0.32, bottomY + 140);
-      ctx.font = "700 36px Inter";
-      ctx.fillStyle = "#FF00FF"; // Pink
-      ctx.fillText("GAMES", width * 0.32, bottomY + 185);
-      ctx.fillText("PLAYED", width * 0.32, bottomY + 225);
-    }
-
-    // Draw days without wedgies or new wedgie (right side)
-    const { daysWithoutWedgie, hasNewWedgie } = getWedgieDays(stats.lastWedgie);
-
-    if (hasNewWedgie) {
-      ctx.font = "900 94px Inter";
-      ctx.fillStyle = "#FF00FF"; // Pink
-      ctx.fillText("NEW", width * 0.68, bottomY + 135);
-      ctx.font = "900 60px Inter";
-      ctx.fillStyle = "#EAFF00"; // Yellow
-      ctx.fillText("WEDGIE", width * 0.68, bottomY + 195);
-    } else {
-      ctx.font = "900 140px Inter";
-      ctx.fillStyle = "#EAFF00"; // Yellow
-      ctx.fillText(daysWithoutWedgie.toString(), width * 0.68, bottomY + 140);
-      ctx.font = "700 36px Inter";
-      ctx.fillStyle = "#FF00FF"; // Pink
-      ctx.fillText("DAYS WITHOUT", width * 0.685, bottomY + 185);
-      ctx.fillText("WEDGIES", width * 0.685, bottomY + 225);
-    }
-
-    drawLogoAndUrlBadge(ctx, width, height, 210, 259);
-  };
-
-  /**
-   * Generates the video by drawing frames onto the canvas.
-   */
-  const generateVideo = async () => {
-    try {
-      setIsGenerating(true);
-      setProgress(0);
-      const canvas = canvasRef.current;
-      if (!canvas) return;
-
-      if (videoType === "desktop") {
-        canvas.width = 1920;
-        canvas.height = 1080;
-      } else {
-        canvas.width = 1080;
-        canvas.height = 1920;
-      }
-
-      const ctx = canvas.getContext("2d");
-      if (!ctx) return;
-
-      // Initialize confetti particles if at 100% fill
-      const fillPercentage = Math.min((stats.totalWedgies / 50) * 100, 100);
-      if (fillPercentage >= 100) {
-        confettiParticlesRef.current = createConfettiParticles(
-          canvas.width,
-          canvas.height,
-        );
-      }
-
-      // Load the logo image
-      const logoUrl =
-        "https://res.cloudinary.com/wedgietracker/image/upload/v1737299676/assets/logo-wedgietracker_rlzejd.png";
-      const logoImage = await loadImage(logoUrl);
-      imageRef.current = logoImage;
-
-      // Check supported MIME types
-      const mimeTypes = [
-        'video/mp4;codecs="avc1.424028, mp4a.40.2"',
-        "video/mp4",
-        "video/webm;codecs=vp9",
-        "video/webm;codecs=vp8",
-        "video/webm",
-      ];
-
-      const supportedMimeType = mimeTypes.find((mimeType) =>
-        MediaRecorder.isTypeSupported(mimeType),
-      );
-
-      if (!supportedMimeType) {
-        throw new Error("No supported media type found");
-      }
-
-      const stream = canvas.captureStream(30); // 30 FPS
-      const mediaRecorder = new MediaRecorder(stream, {
-        mimeType: supportedMimeType,
-        videoBitsPerSecond: 8000000,
-      });
-
-      const chunks: Blob[] = [];
-      mediaRecorder.ondataavailable = (e) => chunks.push(e.data);
-      mediaRecorder.onstop = () => {
-        const blob = new Blob(chunks, { type: supportedMimeType });
-        setVideoBlob(blob);
-        setIsGenerating(false);
-      };
-
-      mediaRecorder.start();
-
-      const startTime = Date.now();
-      const duration = 10000; // 10 seconds
-
-      const animate = () => {
-        const currentProgress = (Date.now() - startTime) / duration;
-        if (currentProgress < 1) {
-          setProgress(currentProgress);
-          if (videoType === "desktop") {
-            drawFrame(ctx, currentProgress, stats);
-          } else {
-            drawVerticalFrame(ctx, currentProgress, stats);
-          }
-          requestAnimationFrame(animate);
-        } else {
-          setProgress(1);
-          mediaRecorder.stop();
-        }
-      };
-
-      animate();
-    } catch (error) {
-      console.error("Error generating video:", error);
-      setIsGenerating(false);
-    }
-  };
-
-  const handleDownload = () => {
-    if (!videoBlob) return;
-
-    const today = new Date();
-    const dateString = today.toISOString().split("T")[0];
-    const url = window.URL.createObjectURL(videoBlob);
-    const a = document.createElement("a");
-    a.href = url;
-
-    // Determine file extension based on MIME type
-    const extension = videoBlob.type.includes("webm") ? "webm" : "mp4";
-    a.download = `WedgieTracker-${dateString}.${extension}`;
-
-    document.body.appendChild(a);
-    a.click();
-    document.body.removeChild(a);
-    window.URL.revokeObjectURL(url);
-  };
+  const {
+    canvasRef,
+    isGenerating,
+    videoBlob,
+    progress,
+    videoType,
+    setVideoType,
+    generateVideo,
+    handleDownload,
+    reset,
+  } = useStatsVideo(stats);
 
   return (
     <Dialog>
@@ -686,40 +71,10 @@ export function ShareableStatsVideo({ stats }: ShareableStatsVideoProps) {
                 </span>
               </button>
               {!isGenerating && (
-                <div className="border-pink flex w-full items-center justify-center gap-4 rounded-full border-2 p-2">
-                  <button
-                    onClick={() => setVideoType("desktop")}
-                    className={`flex-1 rounded-full px-4 py-2 text-sm font-bold transition-all ${
-                      videoType === "desktop"
-                        ? "bg-pink text-darkpurple"
-                        : "text-pink hover:bg-pink/10"
-                    }`}
-                  >
-                    <div className="flex flex-col items-center gap-1">
-                      <RectangleHorizontal className="h-5 w-5" />
-                      <span>Landscape (16:9)</span>
-                      <span className="text-xs opacity-75">
-                        YouTube • Twitter
-                      </span>
-                    </div>
-                  </button>
-                  <button
-                    onClick={() => setVideoType("mobile")}
-                    className={`flex-1 rounded-full px-4 py-2 text-sm font-bold transition-all ${
-                      videoType === "mobile"
-                        ? "bg-pink text-darkpurple"
-                        : "text-pink hover:bg-pink/10"
-                    }`}
-                  >
-                    <div className="flex flex-col items-center gap-1">
-                      <RectangleVertical className="h-5 w-5" />
-                      <span>Portrait (9:16)</span>
-                      <span className="text-xs opacity-75">
-                        Instagram • TikTok
-                      </span>
-                    </div>
-                  </button>
-                </div>
+                <StatsVideoFormatToggle
+                  value={videoType}
+                  onChange={setVideoType}
+                />
               )}
             </>
           )}
@@ -744,34 +99,12 @@ export function ShareableStatsVideo({ stats }: ShareableStatsVideoProps) {
           )}
 
           {videoBlob && (
-            <div className="flex w-full flex-col items-center gap-4">
-              <video
-                src={URL.createObjectURL(videoBlob)}
-                controls
-                className={`w-full rounded-lg ${
-                  videoType === "desktop" ? "aspect-video" : "max-h-[400px]"
-                }`}
-              >
-                <track kind="captions" />
-              </video>
-              <div className="flex w-full flex-row gap-4">
-                <button
-                  onClick={() => {
-                    setVideoBlob(null);
-                    setIsGenerating(false);
-                  }}
-                  className="border-pink text-pink hover:bg-pink hover:text-darkpurple flex-1 rounded-full border-2 bg-transparent px-8 py-2 text-center font-bold tracking-wide uppercase transition-all duration-300"
-                >
-                  Generate New
-                </button>
-                <button
-                  onClick={handleDownload}
-                  className="border-yellow bg-yellow text-darkpurple hover:bg-darkpurple hover:text-yellow flex-1 rounded-full border-2 px-8 py-2 text-center font-black tracking-wide uppercase transition-all duration-300"
-                >
-                  Download
-                </button>
-              </div>
-            </div>
+            <StatsVideoResult
+              videoBlob={videoBlob}
+              videoType={videoType}
+              onReset={reset}
+              onDownload={handleDownload}
+            />
           )}
 
           <div className="mt-2 text-center text-sm text-white/60">
